==> code/Climate Shelter Index/2. NDVI.ipynb <==
This script computes the average NDVI for each green area in the municipality of Bologna

### 0. Libraries

# Libraries
import geopandas as gpd
import rasterio
import numpy as np
from rasterio.features import geometry_mask
from rasterio.warp import calculate_default_transform, reproject, Resampling


### 1. NDVI

# Load the shapefile
green_areas = gpd.read_file(r"D:\Climate_Shelter_Index\ar_vrd_s_df_pt_b.shp")

# Open the GeoTIFF
with rasterio.open("D:/Climate_Shelter_Index/NDVI/ndvi_bologna.tif") as src:
    # Define the target CRS (EPSG:4326)
    target_crs = 'EPSG:4326'

    # Calculate the transformation parameters for reprojection
    transform, width, height = calculate_default_transform(src.crs, target_crs, src.width, src.height, *src.bounds)

    # Create an empty array for the reprojected data
    reprojected_ndvi = np.empty((height, width), dtype=src.dtypes[0])

    # Reproject the data to EPSG:4326
    reproject(
        source=rasterio.band(src, 1),
        destination=reprojected_ndvi,
        src_transform=src.transform,
        src_crs=src.crs,
        dst_transform=transform,
        dst_crs=target_crs,
        resampling=Resampling.bilinear
    )

    # Iterate through green areas
    for index, row in green_areas.iterrows():
        # Get the geometry of the green area
        geometry = row['geometry']
        
        # Create a mask for the reprojected NDVI values based on the green area's geometry
        mask = geometry_mask([geometry], out_shape=reprojected_ndvi.shape, transform=transform, invert=True)
        
        # Extract NDVI values within the green area using the mask
        ndvi_values_within_area = np.ma.masked_array(reprojected_ndvi, mask=~mask)
        
        # Calculate the mean NDVI for the green area while ignoring masked (NaN) values
        mean_ndvi = np.mean(ndvi_values_within_area)
        
        # Update the 'NDVI' column with the computed mean NDVI value
        green_areas.at[index, 'NDVI'] = mean_ndvi

green_areas.head()

# Save the updated shapefile with the 'NDVI' column included
green_areas.to_file('D:/Climate_Shelter_Index/green_areas__ndvi.shp')
