==> code/Climate Shelter Index/4. Bus_stop.ipynb <==
This script identifies bus stops within the isochrones

### 0. Libraries

import geopandas as gpd
import pandas as pd


### 1. Bus stop
Identify the bus stop in each isochrone

# Load the shapefile of bus stops
# link: https://opendata.comune.bologna.it/explore/dataset/tper-fermate-autobus/information/?disjunctive.codice&disjunctive.codice_linea&disjunctive.quartiere
bus_stop = gpd.read_file("D:/Climate_Shelter_Index/fermate_bus/tper-fermate-autobus.shp")
# Load the shapefile of isochrones
isochrones = gpd.read_file ("D:/Climate_Shelter_Index/isocrona/10min_isochrones.shp" )

isochrones

unique_bus_stops = bus_stop.drop_duplicates(subset='codice', keep='first')

# Perform a spatial join to check if each isochrones contains a bus stop
results_df = gpd.sjoin(isochrones[['min_dist','nome', 'geometry']], unique_bus_stops[['codice', 'geometry']], how="left", op="intersects")

# Create a new column 'bus_stop' indicating if there is a bus stop
results_df["bus_stop"] = results_df["codice"].notnull().replace({True: "yes", False: "no"})

# Create a new column 'id_busstop' with the ID of the bus stop if present
results_df["id_busstop"] = results_df["codice"].fillna("N/A")

results_df = results_df[['min_dist','nome', 'codice', 'bus_stop', 'id_busstop']]

results_df = results_df.merge(unique_bus_stops[['codice', 'geometry']], 
                              left_on='id_busstop', right_on='codice', how='left')

# Drop the duplicate  columns and rename 
results_df = results_df.drop(columns=[ 'codice_x', 'codice_y'])

results_df.head()

# Create a GeoDataFrame
results_gdf = gpd.GeoDataFrame(results_df, geometry='geometry', crs="epsg:4326")

# Save the GeoDataFrame as a shapefile
results_gdf.to_file('D:/Climate_Shelter_Index/isocrona/bus_stop_10_ga.shp')

### 2. Number of bus stop per isochrone

bus_stop_ga = gpd.read_file(r"D:\Climate_Shelter_Index\isocrona\bus_stop_10_ga.shp")

# Create a DataFrame to store the aggregation
sum_bus_stop = bus_stop_ga[['nome', 'min_dist']].copy()
# Count the occurrences of each 'nome'
sum_bus_stop['n_busstop'] = sum_bus_stop.groupby('nome')['nome'].transform('count')

sum_bus_stop.head()

# Drop duplicates to keep only unique rows
sum_bus_stop = sum_bus_stop.drop_duplicates(subset=['nome', 'min_dist', 'n_busstop'])

sum_bus_stop.head()

sum_bus_stop.to_csv('D:/Climate_Shelter_Index/isocrona/sum_bus_10_stop.csv', index=False)
