==> code/Climate Shelter Index/5. Disabled_parking.ipynb <==
This code checks whether there is disabled parking within a 5-minute walk of green areas

### 0. Libraries

import geopandas as gpd
import pandas as pd


### 1. Disabled parking
Identify disabled parkings in each isochrone

# Load the shapefile of disabled parkings
# link: https://opendata.comune.bologna.it/explore/dataset/sosta-veicoli-a-servizio-di-persone-disabili-h/information/ 
disab_parking = gpd.read_file("D:/Climate_Shelter_Index/parcheggi_disabili/sosta-veicoli-a-servizio-di-persone-disabili-h.shp")
# Load the shapefile of isochrones
isochrones = gpd.read_file ("D:/Climate_Shelter_Index/isocrona/5min_isochrones.shp" )

disab_parking

results_df =gpd.sjoin(isochrones[['min_dist','nome', 'geometry']], disab_parking[['numero', 'geometry']], 
          how="left", op="intersects")

results_df

# Keep only the desired columns from the filtered DataFrame
results_df = results_df[['nome', 'min_dist', 'numero']]

# Convert 'numero' to integers and fill missing values with 0
results_df['numero'] = results_df['numero'].fillna(0).astype(int)

# Count the occurrences of each 'nome' and sum 'numero'
sum_result = results_df.groupby('nome')['numero'].sum().reset_index()

sum_result = sum_result.rename(columns={'numero':'disab_p_5'})

sum_result

# Save the GeoDataFrame as a shapefile
sum_result.to_csv('D:/Climate_Shelter_Index/isocrona/disabled_parking.csv', index=False)
