==> code/Climate Shelter Index/7. CSI.ipynb <==
This script calculates the Climate Shelter Index for each public green area.  
A climate shelter should have these characteristics:
- Surface area > 0.5 ha 
- NDVI > 0.4 
- Good accessibility
- Presence of drinking fountains 
- Presence of  benches 
- Presence of picnic tables

### 0. Libraries

import geopandas as gpd
import pandas as pd

### 1. Climate Shelter Index - CSI

# Load your shapefile
green_area = gpd.read_file(r'D:\Climate_Shelter_Index\isocrona\access_index')

green_area.head()

green_area.columns

# Define weights for each characteristic (surface area, NDVI, accessibility, drinking fountains, benches, picnic tables)
weights = {
    'surface_area': 0.25,
    'NDVI': 0.25,
    'accessibility': 0.2,
    'drinking_fountains': 0.1,
    'benches': 0.1,
    'picnic_tables': 0.1
}


# Count variables for presence of drinking fountains, benches, and picnic tables
green_area['n_df'] = green_area['id_df'].apply(lambda x: len(x.split(',')) if x != 'N/A' else 0)
green_area['n_bench'] = green_area['id_bench'].apply(lambda x: len(x.split(',')) if x != 'N/A' else 0)
green_area['n_pt'] = green_area['id_pt'].apply(lambda x: len(x.split(',')) if x != 'N/A' else 0)


# Define the columns you want to normalize
columns_to_normalize = ['n_df', 'n_bench', 'n_pt']

# Perform min-max normalization for each column
for col_name in columns_to_normalize:
    min_value = green_area[col_name].min()
    max_value = green_area[col_name].max()
    green_area[col_name] = (green_area[col_name] - min_value) / (max_value - min_value)


# Normalize surface area (0.5 ha threshold)
green_area['norm_area'] = green_area['area_Ha'].apply(lambda x: 1 if x >= 0.5 else 0)


# Normalize NDVI (0.4 threshold)
green_area['norm_NDVI'] = green_area['NDVI'].apply(lambda x: x if x >= 0.4 else 0)


# Calculate the CSI based on the weighted and normalized variables
green_area['CSI'] = (
    weights['surface_area'] * green_area['norm_area'] +
    weights['NDVI'] * green_area['norm_NDVI'] +
    weights['accessibility'] * green_area['access_ind'] +
    weights['drinking_fountains'] * green_area['n_df'] +
    weights['benches'] * green_area['n_bench'] +
    weights['picnic_tables'] * green_area['n_pt']
)

green_area = green_area.drop(columns=['n_df', 'n_bench', 'n_pt', 'norm_area', 'norm_NDVI'])

green_area[green_area['NDVI']<0.41].head()

green_area.to_file(r'D:\Climate_Shelter_Index\isocrona\CSI')

#### 1.2 CSI for selected areas


csi = gpd.read_file(r'D:\Climate_Shelter_Index\isocrona\CSI')
covered_areas = gpd.read_file(r"D:\Meteoblue\covered_areas.shp")

# select only green area within covered areas
covered_areas_csi =csi[csi.geometry.within(covered_areas.unary_union)] 

covered_areas_csi.columns

covered_areas_csi_sorted = covered_areas_csi.sort_values(by='CSI', ascending=False).head(10)

for index, row in covered_areas_csi_sorted.iterrows():
    print(f"{row['nome']}, with CSI of {round(row['CSI'],3)}")

covered_areas_csi_sorted

covered_areas_csi.to_file(r'D:\Climate_Shelter_Index\CSI_covered_areas')

##### 1.2.1 Green Surface > 0.5

covered_areas_csi[covered_areas_csi['area_Ha']>0.5].shape

# find the top 10 parks with the largest surface area
covered_areas_csi_sorted_2 = covered_areas_csi.sort_values(by='area_Ha', ascending=False).head(10)

for index, row in covered_areas_csi_sorted_2.iterrows():
    print(f"{row['nome']}, with a green area of {round(row['area_Ha'], 3)} hectares")

##### 1.2.2 NDVI > 0.4

covered_areas_csi[covered_areas_csi['NDVI']>0.4].shape

##### 1.2.3 NDVI + Green Surface

covered_areas_csi[(covered_areas_csi['NDVI']>0.4) & (covered_areas_csi['area_Ha']>0.5)].shape
